% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_aec_data.R
\name{get_aec_data}
\alias{get_aec_data}
\title{Download and Process AEC Data}
\usage{
get_aec_data(
  file_name,
  date_range = list(from = "2022-01-01", to = "2023-01-01"),
  type = c("Federal Election", "Referendum"),
  category = c("House", "Senate", "General"),
  process = FALSE
)
}
\arguments{
\item{file_name}{The name of the file to download, which is used to construct the download URL
and to filter internal metadata for validations.}

\item{date_range}{A list containing the 'from' and 'to' dates defining the period for which data
is required. Dates should be formatted as 'YYYY-MM-DD'.}

\item{type}{The type of election or event to filter the data, possible values are
'Federal Election' and 'Referendum'.}

\item{category}{The category of the data to be downloaded, possible options are 'House', 'Senate',
and 'General'.}

\item{process}{A logical flag indicating whether additional processing steps (like standardising
column names) should be performed on the downloaded data. Defaults to FALSE.}
}
\value{
A data frame combining all downloaded data, supplemented with relevant information from
internal metadata, and optionally processed for standardisation.
}
\description{
This function is designed to download and process data from the Australian Electoral Commission
(AEC) based on specified criteria. It utilises several internal helper functions to validate
input parameters, construct URLs, and check file availability before downloading and processing the data.
}
\examples{
\dontrun{
get_aec_data(
  file_name = "National list of candidates",
  date_range = list(from = "2022-01-01", to = "2023-01-01"),
  type = "Federal Election",
  category = "House",
  process = FALSE
)
}

}
