% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_disclosure_data.R
\name{get_disclosure_data}
\alias{get_disclosure_data}
\title{Retrieve Disclosure Data from AEC's Transparency Register}
\usage{
get_disclosure_data(
  file_name = NULL,
  group = NULL,
  type = c("Annual", "Election", "Referendum")
)
}
\arguments{
\item{file_name}{The file name of the data to retrieve. If \code{file_name} is not specified, it defaults to
"Donations Made". Must be one of: "Capital Contributions", "Debts", "Discretionary Benefits",
"Donations Made", "Donations Received", "Expenses", "Media Advertisement Details", "Receipts",
"Return Summary", "Returns".}

\item{group}{The group of the entity. If not specified, it defaults to "Donor". Must be one of:
"Associated Entity", "Candidate", "Donor", "Media", "MPs", "Other", "Party", "Referendum Entity",
"Significant Third Party", "Third Party".}

\item{type}{The type of return. Defaults to "Annual". Must be one of: "Annual", "Election", "Referendum".}
}
\value{
A data frame containing the requested disclosure data.
}
\description{
This function downloads and retrieves specific datasets from the Australian Electoral Commission's (AEC)
Transparency Register. The register contains financial disclosure information from political entities,
including annual returns, election returns, and referendum returns.
}
\examples{
\dontrun{
  # Retrieve default data: Donations Made by Donors for Annual returns
  data <- get_disclosure_data()

  # Retrieve specific data: Receipts for Parties in Election returns
  data <- get_disclosure_data(file_name = "Receipts", group = "Party", type = "Election")
}

}
\seealso{
\url{https://www.aec.gov.au/parties_and_representatives/financial_disclosure/} for more
information on the AEC's financial disclosure scheme.
}
